Move utility functions to `.commandline.common`
