+++ test/splitgraph/commandline/snapshots/test_cloud_jobs/test_csv_upload/False/sgr_cloud_upload_failure.txt
@@ -1,5 +1,5 @@
 Uploading the files...
- (STARTED) Loading someuser/somerepo_1, task ID ingest_task
- (FAILURE) Loading someuser/somerepo_1, task ID ingest_task
+ (STARTED) Waiting for task ID ingest_task
+ (FAILURE) Waiting for task ID ingest_task
 
 Logs for /someuser/somerepo_1/ingest_task
